requirements.txt
pytest
mock
mysql-connector-python
